refactor(gcs): Replace complex Closure logic with Callable (#783)
* refactor(gcs): Update retry to new format

This updates the GCS storage backend to consume the new retry
implementation that can return typed objects. This commit does the
minimal change required to get the front50 GCS implementation to work
with the new retry implementation, which is:
(1) Delete the GcsSafeRetry that extends the now final GoogleCommonSafeRetry
and just use a GoogleCommonSafeRetry directly
(2) Replace the Long values in the config with int values (as nobody is waiting
24 days between retries)
(3) Instead of storing all these retry values directly on GcsStorageService,
just create the GoogleCommonSafeRetry that uses them in the constructor
(4) Replace the catch blocks that were expecting some mix of exceptions
to now catch GoogleApiException and return a GcsProviderOperationException
wrapping it with a description of what was happening.

* refactor(gcs): Replace complex Closure logic with Callable

Now that the retry handler can accept a callable and can return
a typed object, replace the extremely complex closure logic in
this file with callables defined as lambdas.

* fix(gcs): Remove extraneous log statement

When we don't find an object, we log then throw a NotFoundException.
In general the benefit of logging then throwing an exception is
questionable; it's better to just include the details on the exception
which will either be handled or logged anyway.

In this case, there are some cases where a NotFoundException is
a reasonable outcome, and we catch it and proceed. In those cases,
the log statement is just adding noise to the logs. Let's remove it
and improve the message on the exception, so that cases that do end
up propagating or logging that exception will still get a good message
but those that choose to ignore the exception don't add noise to
the logs.

* chore(dependencies): Bump clouddriverVersion

+++ front50-gcs/src/main/java/com/netflix/spinnaker/front50/config/GcsConfig.java
@@ -42,16 +42,16 @@ import rx.schedulers.Schedulers;
 @EnableConfigurationProperties(GcsProperties.class)
 public class GcsConfig extends CommonStorageServiceDAOConfig {
   @Value("${spinnaker.gcs.safe-retry.max-wait-interval-ms:60000}")
-  Long maxWaitInterval;
+  int maxWaitInterval;
 
   @Value("${spinnaker.gcs.safe-retry.retry-interval-base-sec:2}")
-  Long retryIntervalBase;
+  int retryIntervalBase;
 
   @Value("${spinnaker.gcs.safe-retry.jitter-multiplier:1000}")
-  Long jitterMultiplier;
+  int jitterMultiplier;
 
   @Value("${spinnaker.gcs.safe-retry.max-retries:10}")
-  Long maxRetries;
+  int maxRetries;
 
   @Value("${spinnaker.gcs.connect-timeout-sec:45}")
   Integer connectTimeoutSec;
